fix: move customRender para pasta utils
