+++ frontend/src/tests/utils/customRender.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import TasksProvider from '../Context/TasksProvider';
+import TasksProvider from '../../Context/TasksProvider';
 
 function customRender(ui) {
   return render(
